+++ src/bump_vec.rs
@@ -306,10 +306,10 @@ impl<T, A: BumpAllocator> BumpVec<T, A> {
             unsafe {
                 if count != 0 {
                     for _ in 0..(count - 1) {
-                        vec.unchecked_push_with(|| value.clone());
+                        vec.push_with_unchecked(|| value.clone());
                     }
 
-                    vec.unchecked_push_with(|| value);
+                    vec.push_with_unchecked(|| value);
                 }
             }
 
@@ -429,7 +429,7 @@ impl<T, A: BumpAllocator> BumpVec<T, A> {
             while vec.len() != vec.capacity() {
                 match iter.next() {
                     // SAFETY: we checked above that `len != capacity`, so there is space
-                    Some(value) => unsafe { vec.unchecked_push(value) },
+                    Some(value) => unsafe { vec.push_unchecked(value) },
                     None => break,
                 }
             }
@@ -707,8 +707,8 @@ impl<T, A: BumpAllocator> BumpVec<T, A> {
     /// # Safety
     /// Vector must not be full.
     #[inline(always)]
-    pub unsafe fn unchecked_push(&mut self, value: T) {
-        unsafe { self.fixed.cook_mut() }.unchecked_push(value);
+    pub unsafe fn push_unchecked(&mut self, value: T) {
+        unsafe { self.fixed.cook_mut() }.push_unchecked(value);
     }
 
     /// Appends an element to the back of the collection.
@@ -716,8 +716,8 @@ impl<T, A: BumpAllocator> BumpVec<T, A> {
     /// # Safety
     /// Vector must not be full.
     #[inline(always)]
-    pub unsafe fn unchecked_push_with(&mut self, f: impl FnOnce() -> T) {
-        unsafe { self.fixed.cook_mut() }.unchecked_push_with(f);
+    pub unsafe fn push_with_unchecked(&mut self, f: impl FnOnce() -> T) {
+        unsafe { self.fixed.cook_mut() }.push_with_unchecked(f);
     }
 
     /// Forces the length of the vector to `new_len`.
@@ -781,7 +781,7 @@ impl<T, A: BumpAllocator> BumpVec<T, A> {
         use fn generic_push_with(&mut self, f: impl FnOnce() -> T) {
             self.generic_reserve_one()?;
             unsafe {
-                self.unchecked_push_with(f);
+                self.push_with_unchecked(f);
             }
             Ok(())
         }
@@ -1032,7 +1032,7 @@ impl<T, A: BumpAllocator> BumpVec<T, A> {
                     let fake = ManuallyDrop::new(MaybeUninit::<T>::uninit().assume_init());
 
                     for _ in 0..count {
-                        self.unchecked_push((*fake).clone());
+                        self.push_unchecked((*fake).clone());
                     }
 
                     return Ok(());