+++ src/fixed_bump_vec.rs
@@ -453,7 +453,7 @@ impl<'a, T> FixedBumpVec<'a, T> {
         use fn generic_push_with(&mut self, f: impl FnOnce() -> T) {
             self.generic_reserve_one()?;
             unsafe {
-                self.unchecked_push_with(f);
+                self.push_with_unchecked(f);
             }
             Ok(())
         }
@@ -711,7 +711,7 @@ impl<'a, T> FixedBumpVec<'a, T> {
                     let fake = ManuallyDrop::new(MaybeUninit::<T>::uninit().assume_init());
 
                     for _ in 0..count {
-                        self.unchecked_push((*fake).clone());
+                        self.push_unchecked((*fake).clone());
                     }
 
                     return Ok(());
@@ -1081,8 +1081,8 @@ impl<'a, T> FixedBumpVec<'a, T> {
     /// # Safety
     /// Vector must not be full.
     #[inline(always)]
-    pub unsafe fn unchecked_push(&mut self, value: T) {
-        self.unchecked_push_with(|| value);
+    pub unsafe fn push_unchecked(&mut self, value: T) {
+        self.push_with_unchecked(|| value);
     }
 
     /// Appends an element to the back of the collection.
@@ -1090,7 +1090,7 @@ impl<'a, T> FixedBumpVec<'a, T> {
     /// # Safety
     /// Vector must not be full.
     #[inline(always)]
-    pub unsafe fn unchecked_push_with(&mut self, f: impl FnOnce() -> T) {
+    pub unsafe fn push_with_unchecked(&mut self, f: impl FnOnce() -> T) {
         debug_assert!(!self.is_full());
 
         let ptr = self.as_mut_ptr().add(self.len());