+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+- **breaking:** renamed `unchecked_push` to `push_unchecked`
+- **breaking:** renamed `unchecked_push_with` to `push_with_unchecked`
+
 ## 0.14.0 (2024-12-12)
 - **breaking:** fix `scoped_aligned`'s closure to take a `BumpScope` with `NEW_MIN_ALIGN` instead of `MIN_ALIGN`
 - **breaking:** `aligned`'s closure now takes a `BumpScope` with lifetime `'a` instead of `'_`